requirements.txt
Flask==2.0.3
Flask_Mail==0.9.1
flask_restx==0.5.1
flask_restful==0.3.9
PyJWT==2.3.0
pymongo==4.1.1
pymongo[srv]
bcrypt
requests==2.25.1
flask-cors
unittest2
pytest==7.1.2
google_auth_oauthlib==0.5.1
numpy
pyfcm